+++ xklib-3d-gl/src/main/java/com/xkball/xklib/x3d/backend/gl/state/BlitRenderState.java
@@ -10,10 +10,10 @@ public record BlitRenderState(
     IRenderPipeline pipeline,
     TextureSetup textureSetup,
     Matrix3x2f pose,
-    int x0,
-    int y0,
-    int x1,
-    int y1,
+    float x0,
+    float y0,
+    float x1,
+    float y1,
     float u0,
     float u1,
     float v0,
@@ -26,10 +26,10 @@ public record BlitRenderState(
         IRenderPipeline pipeline,
         TextureSetup textureSetup,
         Matrix3x2f pose,
-        int x0,
-        int y0,
-        int x1,
-        int y1,
+        float x0,
+        float y0,
+        float x1,
+        float y1,
         float u0,
         float u1,
         float v0,
@@ -48,8 +48,8 @@ public record BlitRenderState(
         vertexConsumer.addVertexWith2DPose(this.pose(), this.x1(), this.y0(), zOffset).setUv(this.u1(), this.v0()).setColor(this.color());
     }
 
-    private static @Nullable ScreenRectangle getBounds(int x0, int y0, int x1, int y1, Matrix3x2f pose, @Nullable ScreenRectangle scissorArea) {
-        ScreenRectangle bounds = new ScreenRectangle(x0, y0, x1 - x0, y1 - y0).transformMaxBounds(pose);
+    private static @Nullable ScreenRectangle getBounds(float x0, float y0, float x1, float y1, Matrix3x2f pose, @Nullable ScreenRectangle scissorArea) {
+        ScreenRectangle bounds = new ScreenRectangle((int) x0, (int) y0, (int) (x1 - x0), (int) (y1 - y0)).transformMaxBounds(pose);
         return scissorArea != null ? scissorArea.intersection(bounds) : bounds;
     }
 }
