+++ xklib-3d/src/main/java/com/xkball/xklib/ui/render/IGUIGraphics.java
@@ -16,26 +16,26 @@ public interface IGUIGraphics {
     
     ITexture getTexture(ResourceLocation location);
     
-    void enableScissor(int x0, int y0, int x1, int y1);
+    void enableScissor(float x0, float y0, float x1, float y1);
     
     void disableScissor();
     
     void submitColoredRectangle(IRenderPipeline pipeline,
-                                int minX, int minY,
-                                int maxX, int maxY,
+                                float minX, float minY,
+                                float maxX, float maxY,
                                 int colorFrom, int colorTo);
     
     void submitColoredRoundedRectangle(
-                                int minX, int minY,
-                                int maxX, int maxY,
-                                int colorFrom, int colorTo, int radius);
+                                float minX, float minY,
+                                float maxX, float maxY,
+                                int colorFrom, int colorTo, float radius);
     
-    void drawString(IFont font, IComponent text, int x, int y, int color, boolean drawShadow);
+    void drawString(IFont font, IComponent text, float x, float y, int color, boolean drawShadow);
     
     void submitBlit(
             IRenderPipeline pipeline,
             ITexture textureView,
-            int x0, int y0, int x1, int y1,
+            float x0, float y0, float x1, float y1,
             float u0, float u1, float v0, float v1,
             int color
     );
@@ -48,34 +48,34 @@ public interface IGUIGraphics {
     
     void layerDown();
     
-    default void renderOutline(int x, int y, int width, int height, int color) {
+    default void renderOutline(float x, float y, float width, float height, int color) {
         this.fill(x, y, x + width, y + 1, color);
         this.fill(x, y + height - 1, x + width, y + height, color);
         this.fill(x, y + 1, x + 1, y + height - 1, color);
         this.fill(x + width - 1, y + 1, x + width, y + height - 1, color);
     }
     
-    default void fillRounded(int minX, int minY, int maxX, int maxY, int color, int radius){
+    default void fillRounded(float minX, float minY, float maxX, float maxY, int color, float radius){
         this.submitColoredRoundedRectangle(minX, minY, maxX, maxY, color, color, radius);
     }
     
-    default void fillGradientRounded(int minX, int minY, int maxX, int maxY, int colorFrom, int colorTo, int radius){
+    default void fillGradientRounded(float minX, float minY, float maxX, float maxY, int colorFrom, int colorTo, float radius){
         this.submitColoredRoundedRectangle(minX, minY, maxX, maxY, colorFrom, colorTo, radius);
     }
     
-    default void fill(int minX, int minY, int maxX, int maxY, int color) {
+    default void fill(float minX, float minY, float maxX, float maxY, int color) {
         this.fill(IRenderPipelineSource.getInstance().getGui(), minX, minY, maxX, maxY, color);
     }
     
-    default void fill(IRenderPipeline pipeline, int minX, int minY, int maxX, int maxY, int color) {
+    default void fill(IRenderPipeline pipeline, float minX, float minY, float maxX, float maxY, int color) {
         if (minX < maxX) {
-            int i = minX;
+            float i = minX;
             minX = maxX;
             maxX = i;
         }
         
         if (minY < maxY) {
-            int j = minY;
+            float j = minY;
             minY = maxY;
             maxY = j;
         }
@@ -83,17 +83,17 @@ public interface IGUIGraphics {
         this.submitColoredRectangle(pipeline, minX, minY, maxX, maxY, color, color);
     }
     
-    default void fill(IRenderPipeline pipeline, int minX, int minY, int maxX, int maxY) {
+    default void fill(IRenderPipeline pipeline, float minX, float minY, float maxX, float maxY) {
         this.submitColoredRectangle(pipeline, minX, minY, maxX, maxY, -1, -1);
     }
     
-    default void fillGradient(int minX, int minY, int maxX, int maxY, int colorFrom, int colorTo) {
+    default void fillGradient(float minX, float minY, float maxX, float maxY, int colorFrom, int colorTo) {
         this.submitColoredRectangle(IRenderPipelineSource.getInstance().getGui(), minX, minY, maxX, maxY, colorFrom, colorTo);
     }
     
-    default void hLine(int minX, int maxX, int y, int color) {
+    default void hLine(float minX, float maxX, float y, int color) {
         if (maxX < minX) {
-            int i = minX;
+            float i = minX;
             minX = maxX;
             maxX = i;
         }
@@ -102,9 +102,9 @@ public interface IGUIGraphics {
     }
     
 
-    default void vLine(int x, int minY, int maxY, int color) {
+    default void vLine(float x, float minY, float maxY, int color) {
         if (maxY < minY) {
-            int i = minY;
+            float i = minY;
             minY = maxY;
             maxY = i;
         }
@@ -112,58 +112,58 @@ public interface IGUIGraphics {
         this.fill(x, minY + 1, x + 1, maxY, color);
     }
     
-    default void drawCenteredString(IFont font, String text, int x, int y, int color) {
+    default void drawCenteredString(IFont font, String text, float x, float y, int color) {
         this.drawString(font, text, x - font.width(text) / 2, y, color);
     }
     
-    default void drawCenteredString(String text, int x, int y, int color) {
+    default void drawCenteredString(String text, float x, float y, int color) {
         var font = this.defaultFont();
         this.drawString(font, text, x - font.width(text) / 2, y, color);
     }
     
-    default void drawCenteredString(String text, int x, int y, int color, int height) {
+    default void drawCenteredString(String text, float x, float y, int color, float height) {
         var font = this.defaultFont();
         var scale = height / (float)font.lineHeight();
-        this.drawString(font, text, (int) (x - font.width(text) * scale / 2), y, color, height);
+        this.drawString(font, text, x - font.width(text) * scale / 2, y, color, height);
     }
     
-    default void drawString(IFont font, @Nullable String text, int x, int y, int color, int height){
+    default void drawString(IFont font, @Nullable String text, float x, float y, int color, float height){
         this.getPose().pushMatrix();
         var scale = height / (float)font.lineHeight();
         this.getPose().scale(scale,scale);
-        this.drawString(font, text, (int) (x/scale), (int) (y/scale), color);
+        this.drawString(font, text, x/scale, y/scale, color);
         this.getPose().popMatrix();
     }
     
-    default void drawString(@Nullable String text, int x, int y, int color, int height){
+    default void drawString(@Nullable String text, float x, float y, int color, float height){
         this.drawString(this.defaultFont(), text, x, y, color, height);
     }
     
-    default void drawString(@Nullable String text, int x, int y, int color) {
+    default void drawString(@Nullable String text, float x, float y, int color) {
         this.drawString(this.defaultFont(), text, x, y, color, false);
     }
     
-    default void drawString(IFont font, @Nullable String text, int x, int y, int color) {
+    default void drawString(IFont font, @Nullable String text, float x, float y, int color) {
         this.drawString(font, text, x, y, color, false);
     }
     
-    default void drawString(IFont font, @Nullable String text, int x, int y, int color, boolean drawShadow) {
+    default void drawString(IFont font, @Nullable String text, float x, float y, int color, boolean drawShadow) {
         if (text != null) {
             this.drawString(font, IComponent.literal(text), x, y, color, drawShadow);
         }
     }
     
-    default void drawString(IFont font, IComponent text, int x, int y, int color) {
+    default void drawString(IFont font, IComponent text, float x, float y, int color) {
         this.drawString(font, text, x, y, color, false);
     }
     
-    default void blitSprite(IRenderPipeline renderPipeline, ResourceLocation location, int x, int y, int width, int height, int color) {
+    default void blitSprite(IRenderPipeline renderPipeline, ResourceLocation location, float x, float y, float width, float height, int color) {
         ITextureAtlasSprite sprite = this.getSprite(location);
         this.blitSprite(renderPipeline, sprite, x, y, width, height, color);
     }
     
     default void blitSprite(
-            IRenderPipeline renderPipeline, ResourceLocation location, int spriteWidth, int spriteHeight, int textureX, int textureY, int x, int y, int width, int height
+            IRenderPipeline renderPipeline, ResourceLocation location, float spriteWidth, float spriteHeight, float textureX, float textureY, float x, float y, float width, float height
     ) {
         this.blitSprite(renderPipeline, location, spriteWidth, spriteHeight, textureX, textureY, x, y, width, height, -1);
     }
@@ -171,14 +171,14 @@ public interface IGUIGraphics {
     default void blitSprite(
             IRenderPipeline renderPipeline,
             ResourceLocation location,
-            int spriteWidth,
-            int spriteHeight,
-            int textureX,
-            int textureY,
-            int x,
-            int y,
-            int width,
-            int height,
+            float spriteWidth,
+            float spriteHeight,
+            float textureX,
+            float textureY,
+            float x,
+            float y,
+            float width,
+            float height,
             int color
     ) {
         ITextureAtlasSprite sprite = this.getSprite(location);
@@ -187,11 +187,11 @@ public interface IGUIGraphics {
         this.disableScissor();
     }
     
-    default void blitSprite(IRenderPipeline renderPipeline, ITextureAtlasSprite sprite, int x, int y, int width, int height) {
+    default void blitSprite(IRenderPipeline renderPipeline, ITextureAtlasSprite sprite, float x, float y, float width, float height) {
         this.blitSprite(renderPipeline, sprite, x, y, width, height, -1);
     }
     
-    default void blitSprite(IRenderPipeline renderPipeline, ITextureAtlasSprite sprite, int x, int y, int width, int height, int color) {
+    default void blitSprite(IRenderPipeline renderPipeline, ITextureAtlasSprite sprite, float x, float y, float width, float height, int color) {
         if (width != 0 && height != 0) {
             this.innerBlit(
                     renderPipeline, sprite.atlasLocation(), x, x + width, y, y + height, sprite.getU0(), sprite.getU1(), sprite.getV0(), sprite.getV1(), color
@@ -202,14 +202,14 @@ public interface IGUIGraphics {
     private void blitSprite(
             IRenderPipeline renderPipeline,
             ITextureAtlasSprite sprite,
-            int spriteWidth,
-            int spriteHeight,
-            int textureX,
-            int textureY,
-            int x,
-            int y,
-            int width,
-            int height,
+            float spriteWidth,
+            float spriteHeight,
+            float textureX,
+            float textureY,
+            float x,
+            float y,
+            float width,
+            float height,
             int color
     ) {
         if (width != 0 && height != 0) {
@@ -232,14 +232,14 @@ public interface IGUIGraphics {
     default void blit(
             IRenderPipeline renderPipeline,
             ResourceLocation texture,
-            int x,
-            int y,
+            float x,
+            float y,
             float u,
             float v,
-            int width,
-            int height,
-            int textureWidth,
-            int textureHeight,
+            float width,
+            float height,
+            float textureWidth,
+            float textureHeight,
             int color
     ) {
         this.blit(renderPipeline, texture, x, y, u, v, width, height, width, height, textureWidth, textureHeight, color);
@@ -254,16 +254,16 @@ public interface IGUIGraphics {
     default void blit(
             IRenderPipeline renderPipeline,
             ResourceLocation texture,
-            int x,
-            int y,
+            float x,
+            float y,
             float u,
             float v,
-            int width,
-            int height,
-            int srcWidth,
-            int srcHeight,
-            int textureWidth,
-            int textureHeight
+            float width,
+            float height,
+            float srcWidth,
+            float srcHeight,
+            float textureWidth,
+            float textureHeight
     ) {
         this.blit(renderPipeline, texture, x, y, u, v, width, height, srcWidth, srcHeight, textureWidth, textureHeight, -1);
     }
@@ -271,16 +271,16 @@ public interface IGUIGraphics {
     default void blit(
             IRenderPipeline renderPipeline,
             ResourceLocation texture,
-            int x,
-            int y,
+            float x,
+            float y,
             float u,
             float v,
-            int width,
-            int height,
-            int srcWidth,
-            int srcHeight,
-            int textureWidth,
-            int textureHeight,
+            float width,
+            float height,
+            float srcWidth,
+            float srcHeight,
+            float textureWidth,
+            float textureHeight,
             int color
     ) {
         this.innerBlit(
@@ -298,12 +298,12 @@ public interface IGUIGraphics {
         );
     }
     
-    default void blit(ResourceLocation location, int x0, int y0, int x1, int y1, float u0, float u1, float v0, float v1) {
+    default void blit(ResourceLocation location, float x0, float y0, float x1, float y1, float u0, float u1, float v0, float v1) {
         this.innerBlit(IRenderPipelineSource.getInstance().getGuiTextured(), location, x0, x1, y0, y1, u0, u1, v0, v1, -1);
     }
     
     default void innerBlit(
-            IRenderPipeline renderPipeline, ResourceLocation location, int x0, int x1, int y0, int y1, float u0, float u1, float v0, float v1, int color
+            IRenderPipeline renderPipeline, ResourceLocation location, float x0, float x1, float y0, float y1, float u0, float u1, float v0, float v1, int color
     ) {
         ITexture texture = this.getTexture(location);
         this.submitBlit(renderPipeline, texture, x0, y0, x1, y1, u0, u1, v0, v1, color);
